post himanshu update
this update is directly after himanshu helped debug the "convergence issue. This issue plauged me for months where the action would converge to the most extreme values once the policy was used in the action generation. After much debugging the observation space operated in a range of -1 to 1 much better and seemed to solve the issue.